feat: add security sentinel env config
