+++ .env.example
@@ -32,6 +32,9 @@ TOOLSERVER_URL=http://localhost:8080
 SEC_SENTINEL_ENABLED=true
 SEC_SENTINEL_URL=
 SEC_SENTINEL_API_KEY=
+SEC_SENTINEL_TIMEOUT_MS=4000
+SEC_SENTINEL_RETRIES=1
+SEC_SENTINEL_MODEL=
 SEC_SENTINEL_FAIL_MODE=fallback   # fallback | allow | deny
 
 # Paths and miscellaneous
