+++ libs/GatewayPlugins/SecuritySentinelGatewayPlugin/SecuritySentinelGatewayPlugin/LLMSecuritySentinelClient.swift
@@ -7,17 +7,6 @@ enum FailMode: String {
     case allow, deny, fallback
 }
 
-/// Environment configuration for the LLM Security Sentinel client.
-enum SentinelEnv {
-    static let enabled = (ProcessInfo.processInfo.environment["SEC_SENTINEL_ENABLED"] ?? "true").lowercased() != "false"
-    static let url = ProcessInfo.processInfo.environment["SEC_SENTINEL_URL"]
-    static let apiKey = ProcessInfo.processInfo.environment["SEC_SENTINEL_API_KEY"]
-    static let timeoutMS = Int(ProcessInfo.processInfo.environment["SEC_SENTINEL_TIMEOUT_MS"] ?? "4000") ?? 4000
-    static let retries = Int(ProcessInfo.processInfo.environment["SEC_SENTINEL_RETRIES"] ?? "1") ?? 1
-    static let model = ProcessInfo.processInfo.environment["SEC_SENTINEL_MODEL"]
-    static let failMode = ProcessInfo.processInfo.environment["SEC_SENTINEL_FAIL_MODE"] ?? "fallback"
-}
-
 /// Client that consults an external LLM backed Security Sentinel service.
 final class LLMSecuritySentinelClient: SecuritySentinelClient {
     private let http: HTTPClient
